--- src/daf_gui/tk_object_define/__init__.py
@@ -1,7 +1,0 @@
-"""
-Tk-Object-Define is a library built on top of Tkinter // TTKbootstrap, that enables
-users to create abstract objects based on class's parameter annotations (type hints).
-These can then be convert to real Python objects (or from), with the help of a few utility functions.
-"""
-
-__version__ = "1.0.0"
