+++ src/daf_gui/tkclasswiz/object_frame/frame_string.py
