+++ src/daf_gui/tod_extensions/help_button.py
@@ -1,7 +1,7 @@
 """
 Help button TOD extension.
 """
-from ..tk_object_define.object_frame.frame_struct import NewObjectFrameStruct
+from ..tkclasswiz.object_frame.frame_struct import NewObjectFrameStruct
 
 import ttkbootstrap as ttk
 import webbrowser
