+++ src/daf_gui/tod_extensions/live_objects.py
@@ -2,7 +2,7 @@
 Implements live object tracking to ObjectInfo.
 """
 from collections.abc import Mapping
-from ..tk_object_define.convert import ObjectInfo
+from ..tkclasswiz.convert import ObjectInfo
 from daf.misc.instance_track import ObjectReference
